fix: move function to repo file
